+++ Assets/Scripts/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c3bfd621ccbae7f44b49b86c0d6a83c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
