+++ Assets/Scripts/Entities/Creature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 317d918af9ba07342a4f74ca85c18e95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
