+++ Assets/Scripts/Entities/PlathformThrough.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[RequireComponent(typeof(PlatformEffector2D))]
+public class PlathformThrough : MonoBehaviour
+{
+    private Collider2D coll;
+    private bool playerOnPlatform;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
