+++ Assets/Scripts/Entities/PlathformThrough.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2d57fd89b766f5341a23bcbd674b4803
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
