+++ Assets/Scripts/Entities/Trap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2ed3e21bf386f24799bcc10ae1b8ba3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
